demo-projects/project-04-enterprise/applications/admin-panel/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
psycopg2-binary==2.9.7
redis==4.6.0
python-dotenv==1.0.0
Werkzeug==2.3.6
gunicorn==21.2.0
requests==2.31.0
bcrypt==4.0.1
